[chore] moved navbar component files to separate folders
