+++ src/app/globals.css
@@ -70,7 +70,7 @@
 
    .glass {
       background-color: rgba(255, 255, 255, 0.5);
-      backdrop-filter: blur(20px);
+      backdrop-filter: blur(10px);
       box-shadow: 0px 3px 3px 0px rgba(0, 0, 0, 0.15), 
                   inset 0px 2px 2px 0px rgba(255, 255, 255, 0.5);
    }
