+++ src/components/navbar/topnavbar/ProfileButton.tsx
@@ -19,20 +19,20 @@ export default function ProfileButton({initialOpen = false}: Props) {
    
    return (
       <div ref={ref} className="self-center">
-         <button className="group glass flex w-54 h-10 space-x-2 items-center rounded-4xl pl-6 hover:not-active:scale-105 transition-all active:brightness-75 cursor-pointer"
+         <button className="group glass flex w-20 md:w-54 h-10 space-x-2 items-center rounded-4xl pl-2 md:pl-6 hover:not-active:scale-105 transition-all active:brightness-75 cursor-pointer"
             onClick={() => setIsOpen(!isOpen)}
          >
             {isOpen ? (
                <ChevronUp 
                   size={27} 
                   strokeWidth={2.5} 
-                  className="mt-1 blur-[0.75px] group-hover:not-group-active:stroke-[#FF8829]"
+                  className="mt-1 blur-[0.75px] group-hover:not-group-active:stroke-[#FF8829] hidden md:inline-block"
                />
             ):(
                <ChevronDown 
                   size={27} 
                   strokeWidth={2.5} 
-                  className="mt-1 blur-[0.75px] group-hover:not-group-active:stroke-[#FF8829]"
+                  className="mt-1 blur-[0.75px] group-hover:not-group-active:stroke-[#FF8829] hidden md:inline-block"
                />
             )}
 
@@ -45,10 +45,11 @@ export default function ProfileButton({initialOpen = false}: Props) {
                   className="size-7 rounded-[50%]"
                />
             </div>
-            <span className="blur-[0.75px] text-xl group-hover:not-group-active:text-[#FF8829]">julia apples</span>
+            <span className="blur-[0.75px] text-xl group-hover:not-group-active:text-[#FF8829] hidden md:inline-block">
+               julia apples
+            </span>
          </button>
 
-         {/* {isOpen && <ServiceDropDown menuItems={menuItems} />} */}
          {isOpen && <ProfileDropDown menuItems={menuItems}/>}
       </div>
    );
